# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Backend",
        "version": "0.1"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "summary": "Авторизация",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["login", "password"],
                                "properties": {
                                    "login": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {"$ref": "#/components/responses/Context"}
                }
            }
        },
...
        "/api/grant/user/{userId}/{entityType}/{entityId}": {
            "post": {
                "summary": "Назначение прав доступа",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "schema": {"type": "integer"}
                    },
                    {"$ref": "#/components/parameters/entityType"},
                    {"$ref": "#/components/parameters/entityId"}
                ],
                "requestBody": {
                    "description": "Права доступа",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "create": {"type": "boolean"},
                                    "read": {"type": "boolean"},
                                    "update": {"type": "boolean"},
                                    "delete": {"type": "boolean"},
                                    "grant": {"type": "boolean"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "parameters": {
            "entityId": {
                "name": "entityId",
                "in": "path",
                "schema": {"type": "integer"}
            },
...
            "entityType": {
                "name": "entityType",
                "in": "path",
                "schema": {
                    "type": "string",
                    "enum": [
                        "author",
                        "book",
                        "book_comment",
                        "book_tag",
                        "chapter",
                        "genre",
                        "team",
                        "team_comment",
                        "user"
                    ]
                }
            }
        },
        "responses": {
            "Context": {
                "description": "Контекст клиента",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "context": {
                                    "$ref": "#/components/schemas/Context"
                                }
                            }
                        }
                    }
                }
            }
        },
        "schemas": {
...
            "Context": {
                "type": "object",
                "properties": {
                    "user": {
                        "oneOf": [
                            {"type": "null"},
                            {"$ref": "#/components/schemas/User"}
                        ]
                    }
                }
            },
            "Entity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    }
                }
            },
...
            "User": {
                "allOf": [
                    {"$ref": "#/components/schemas/Entity"},
                    {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "string"
                            }
                        }
                    }
                ]
            },
...
        }
    }
}